tests/test_main_app.py
import pytest
from main_app import MainApp
import streamlit as st

def test_main_app_initialization():
    """Test the initialization of the MainApp component."""
    assert True  # Placeholder for actual test logic

def test_main_app_navigation():
    """Test the navigation functionality of the MainApp component."""
    assert True  # Placeholder for actual test logic
def test_initialize_session_state():
    app = MainApp()
    app.initialize_session_state()
    assert 'project' in st.session_state
    assert 'dark_mode' in st.session_state

def test_setup_theme():
    app = MainApp()
    app.setup_theme()
    # Assertions would depend on how you can check `st.set_page_config`

def test_create_sidebar():
    app = MainApp()
    selected_page = app.create_sidebar()
    # Since Streamlit doesn't return values, this might require integration tests
